# teiv/src/main/resources/v1/topology-exposure-inventory-openapi.yaml
...
openapi: 3.0.3
info:
  x-api-id: 52812f69-83ac-4dfa-b83a-1a1bdf2d49b8
  x-audience: external-public
  description: |
    Topology Exposure and Inventory data is the information that represents entities
    in a telecommunications network and the relationships between them that
    provide insight into a particular aspect of the network of importance to
    specific use cases. Topology and Inventory data can be derived from
    inventory, configuration, or other data.

    Topology Exposure and Inventory supports several topology domains. A domain is a
    grouping of topology and inventory entities that handles topology and
    inventory data.

    Entities are enabling the modelling and storage of complex network
    infrastructure and relationships.

    A relationship is a bi-directional connection between two entities, one
    of which is the originating side (A-side) and the other is the
    terminating side (B-side). The order of the sides matters since it
    defines the relationship itself which must be unique.

    Classifier (also known as tag or label) permits the association of a 
    well defined user specified string with an entity or relationship.

    Decorators are user-defined attributes (key-value pairs) which can
    be applied to topology entities and relationships.

    Metadata provides additional information about entities and relationships within the database.
...
    the first time in Topology & Inventory. 
...

...

...
    query requests.

...

...

...

...

...

...

...

...

...

...
    
...

...

...
    
...

...

...

...
servers:
...
paths:
...
  /schemas:
    post:
      deprecated: true
      description: Create a new schema. The request body contains the schema in YANG format.
      tags:
        - Schemas
      summary: Create a new schema.
      operationId: createSchema
      parameters:
        - $ref: '#/components/parameters/acceptJsonInHeader'
        - $ref: '#/components/parameters/contentTypeMultipartFileInHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultipartFile'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      description: Get a list of all topology model schemas.
      tags:
        - Schemas
      summary: Get a list of all pre defined schemas.
      operationId: getSchemas
      parameters:
        - $ref: '#/components/parameters/acceptJsonInHeader'
        - $ref: '#/components/parameters/domainOptionalInQuery'
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schemas'
              examples:
                schemas:
                  $ref: '#/components/examples/SchemasResponseExample'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '406':
          $ref: '#/components/responses/406'
        '414':
          $ref: '#/components/responses/414'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
  /schemas/{schemaName}/content:
    get:
      description: Get the model schema by name.
      tags:
        - Schemas
      summary: Get the model schema.
      operationId: getSchemaByName
      parameters:
        - $ref: '#/components/parameters/acceptPlainTextInHeader'
        - $ref: '#/components/parameters/schemaNameInPath'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              examples:
                schema:
                  $ref: '#/components/examples/SchemaResponseExample'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '406':
          $ref: '#/components/responses/406'
        '414':
          $ref: '#/components/responses/414'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
  /user-defined-schemas:
    post:
      description: Create a new user defined schema. The request body contains the schema in YANG format.
      tags:
        - User Defined Schemas
      summary: Create a new user defined schema.
      operationId: createUserDefinedSchema
      parameters:
        - $ref: '#/components/parameters/acceptJsonInHeader'
        - $ref: '#/components/parameters/contentTypeMultipartFileInHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultipartFile'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDefinedSchema'
          headers:
            Location:
              description: 'Contains the URI of the newly created resource'
              required: true
              schema:
                type: string
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '409':
          $ref: '#/components/responses/409'
        '411':
          $ref: '#/components/responses/411'
        '413':
          $ref: '#/components/responses/413'
        '415':
          $ref: '#/components/responses/415'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'

    get:
      description: Get a list of all user defined schemas.
      tags:
        - User Defined Schemas
      summary: Get a list of all used defined schemas.
      operationId: getUserDefinedSchemas
      parameters:
        - $ref: '#/components/parameters/acceptJsonInHeader'
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDefinedSchemas'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '406':
          $ref: '#/components/responses/406'
        '414':
          $ref: '#/components/responses/414'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
  /user-defined-schemas/{schemaName}/content:
    get:
      description: Get the user defined model schema by name.
      tags:
        - User Defined Schemas
      summary: Get the user defined model schema.
      operationId: getUserDefinedSchemaByName
      parameters:
        - $ref: '#/components/parameters/acceptPlainTextInHeader'
        - $ref: '#/components/parameters/schemaNameInPath'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              examples:
                schema:
                  $ref: '#/components/examples/SchemaResponseExample'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '406':
          $ref: '#/components/responses/406'
        '414':
          $ref: '#/components/responses/414'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
...
  /user-defined-schemas/{schemaName}:
    delete:
      description: Delete a user defined schema.
      tags:
        - User Defined Schemas
      summary: Delete a user defined schema.
      operationId: deleteUserDefinedSchema
      parameters:
        - $ref: '#/components/parameters/acceptJsonInHeader'
        - $ref: '#/components/parameters/schemaNameInPath'
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
        '502':
          $ref: '#/components/responses/502'
        '503':
          $ref: '#/components/responses/503'
...
components:
  schemas:
...
    ProblemDetails:
      description: 'A problem detail to carry details in an HTTP response according to RFC 7807'
      type: object
      properties:
        type:
          description: 'a URI reference according to IETF RFC 3986 that identifies the problem type'
          type: string
        title:
          description: 'human-readable summary of the problem type'
          type: string
        status:
          description: 'the HTTP status code'
          type: number
        detail:
          description: 'human-readable explanation '
          type: string
        instance:
          description: 'URI reference that identifies the specific occurrence of the problem'
          type: string
    Href:
      type: object
      title: Href
      properties:
        href:
          type: string
          format: uri-template
    MultipartFile:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          description: multipartFile
          format: binary
    Schema:
      type: object
      title: Schema
      properties:
        name:
          type: string
        domain:
          type: string
        revision:
          type: string
        content:
          $ref: '#/components/schemas/Href'
    Schemas:
      type: object
      title: Schemas
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Schema'
    UserDefinedSchema:
      type: object
      title: UserDefinedSchema
      properties:
        name:
          type: string
        domain:
          type: string
        revision:
          type: string
        content:
          $ref: '#/components/schemas/Href'
    UserDefinedSchemas:
      type: object
      title: UserDefinedSchemas
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDefinedSchema'
...
  responses:
    '204':
      description: 'No Content'
      content: { }
    '400':
      description: 'Bad Request'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '401':
      description: 'Unauthorized'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '403':
      description: 'Forbidden'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '404':
      description: 'Not Found'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    '406':
      description: 'Not Acceptable'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '409':
      description: 'Conflict'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '411':
      description: 'Length Required'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '413':
      description: 'Payload Too Large'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '414':
      description: 'URI Too Large'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '415':
      description: 'Unsupported Media Type'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '429':
      description: 'Too Many Requests'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '500':
      description: 'Internal Server Error'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '502':
      description: 'Bad Gateway'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    '503':
      description: 'Service Unavailable'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  parameters:
    acceptJsonInHeader:
      name: Accept
      in: header
      required: true
      schema:
        type: string
        example: application/json
        default: application/json
...
    acceptPlainTextInHeader:
      name: Accept
      in: header
      required: true
      schema:
        type: string
        default: text/plain
    contentTypeMultipartFileInHeader:
      name: Content-Type
      in: header
      required: true
      schema:
        type: string
        example: multipart/form-data
        default: multipart/form-data
...
    offsetParam:
      name: offset
      in: query
      description: Pagination offset.
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    limitParam:
      name: limit
      in: query
      description: Result limiter.
      required: false
      schema:
        type: integer
        default: 500
        minimum: 1
        maximum: 500
...
    schemaNameInPath:
      name: schemaName
      in: path
      required: true
      schema:
        type: string
        default: "o-ran-smo-teiv-ran"
...
    domainOptionalInQuery:
      name: domain
      in: query
      required: false
      schema:
        type: string
      examples:
        domain:
          value: RAN
...
  examples:
...
    SchemasResponseExample:
      value:
        items:
          - name: "o-ran-smo-teiv-ran"
            domain: "RAN"
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-ran/content"
          - name: "o-ran-smo-teiv-equipment"
            domain: "EQUIPMENT"
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-equipment/content"
          - name: "o-ran-smo-teiv-oam"
            domain: "OAM"
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-oam/content"
          - name: "o-ran-smo-teiv-rel-oam-ran"
            domain: "REL_OAM_RAN"
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-rel-oam-ran/content"
          - name: "o-ran-smo-teiv-rel-equipment-ran"
            domain: "REL_EQUIPMENT_RAN"
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-rel-equipment-ran/content"
          - name: "o-ran-smo-teiv-common-yang-types"
            domain: ""
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-common-yang-types/content"
          - name: "o-ran-smo-teiv-common-yang-extensions"
            domain: ""
            revision: "2024-05-24"
            content:
              href: "/schemas/o-ran-smo-teiv-common-yang-extensions/content"
          - name: "ietf-geo-location"
            domain: ""
            revision: "2022-02-11"
            content:
              href: "/schemas/ietf-geo-location/content"
          - name: "_3gpp-common-yang-extensions"
            domain: ""
            revision: "2019-06-23"
            content:
              href: "/schemas/_3gpp-common-yang-extensions/content"
          - name: "_3gpp-common-yang-types"
            domain: ""
            revision: "2023-11-06"
            content:
              href: "/schemas/_3gpp-common-yang-types/content"
          - name: "ietf-yang-types"
            domain: ""
            revision: "2013-07-15"
            content:
              href: "/schemas/ietf-yang-types/content"
          - name: "ietf-inet-types"
            domain: ""
            revision: "2013-07-15"
            content:
              href: "/schemas/ietf-inet-types/content"
        self:
          href: "/schemas?offset=0&limit=500"
        first:
          href: "/schemas?offset=0&limit=500"
        prev:
          href: "/schemas?offset=0&limit=500"
        next:
          href: "/schemas?offset=0&limit=500"
        last:
          href: "/schemas?offset=0&limit=500"
        totalCount: 12
    SchemaResponseExample:
      value: |
        module o-ran-smo-teiv-ran {
        yang-version 1.1;
        namespace "urn:o-ran:smo-teiv-ran";
        prefix or-teiv-ran;
        import o-ran-smo-teiv-common-yang-types {prefix or-teiv-types; }
        import o-ran-smo-teiv-common-yang-extensions {prefix or-teiv-yext; }
        import _3gpp-common-yang-types { prefix types3gpp; }
        organization "ORAN";
        description
        "Sample Model.";
        revision "2024-05-24" {
          description "Initial revision."
          or-teiv-yext:label 0.1.0;
        }
        or-teiv-yext:domain RAN;
        list ODUFunction {
          uses or-teiv-types:Top_Grp_Type;
          key id;
          container attributes {
            container dUpLMNId {
              uses types3gpp:PLMNId;
            }
            leaf gNBDUId {
              type uint32;
            }
            leaf gNBId {
              type uint32;
            }
            leaf gNBIdLength {
              type uint32;
            }
          }
        }
        list NRCellDU {
          uses or-teiv-types:Top_Grp_Type;
          key id;
          container attributes {
            leaf cellLocalId {
              type uint32;
            }
            leaf nCI {
              type uint32;
            }
            leaf nRPCI {
              type uint32;
            }
            leaf nRTAC {
              type uint32;
            }
          }
        }
        list NRSectorCarrier {
          uses or-teiv-types:Top_Grp_Type;
          key id;
            container attributes {
              leaf arfcnDL {
                type uint32;
              }
              leaf arfcnUL {
                type uint32;
              }
              leaf frequencyDL {
                type uint32;
              }
              leaf frequencyUL {
                type uint32;
              }
              leaf bSChannelBwDL {
                type uint32;
              }
            }
          }
        }
...